Connect MySQL and Change id's type to Integer

+++ src/main/java/us/repository/ParticipantRepository.java
@@ -3,8 +3,6 @@ package us.repository;
 import org.springframework.data.repository.CrudRepository;
 import us.model.Participant;
 
-/**
- * Created by PrinceY on 2017-03-13.
- */
-public interface ParticipantRepository extends CrudRepository<Participant, Long> {
+
+public interface ParticipantRepository extends CrudRepository<Participant, Integer> {
 }
